refactor: decompose cli.ts (#126)
Split four mixed concerns out of `src/cli.ts` (3,462 → 2,401 lines, ~1,061
lines extracted) into four focused, unit-testable modules:

- `src/output-context.ts` — `OutputMode` type plus an `initOutputMode()` /
  `getOutputMode()` process-level singleton. Output mode is now resolved
  once at startup from the (normalized) argv and persisted config; the
  global `output()` helper reads the in-memory singleton instead of
  re-scanning `process.argv` and re-calling `loadConfig()` on every call.
  `resolveOutputMode()` becomes a pure function over a synthetic argv
  array and config defaults.
- `src/output-shapes.ts` — `shapeForCommand` and friends (search /
  registry-search / show shaping, agent-mode field picking, description
  truncation). Pure functions over plain objects.
- `src/output-text.ts` — `formatPlain` switch plus all per-command
  plain-text formatters (search, curate, show, wiki-*, workflow-*, list,
  add, remove, update, upgrade, clone) and the `outputJsonl` helper.
- `src/curate.ts` — `akmCurate()` public API plus `curateSearchResults`,
  `orderCuratedTypes`, `deriveCurateFallbackQueries`,
  `mergeCurateSearchResponses`, and `searchForCuration`. The CLI
  `curate` command now delegates to a single `akmCurate({ query, type,
  limit, source })` call. Tests can pass a pre-built `SearchResponse`
  via `searchResponse:` to curate fixtures without IO.

Also fixes the manual `parseFlagValue` argv-scanning helper used by
`saveCommand`'s positional/flag disambiguation: it now takes argv as
an explicit parameter (re-exported from `output-context.ts`) instead
of reading `process.argv` from module scope.

No behavioral changes intended. All 27 `akm <subcommand>` definitions
remain in `src/cli.ts`; this is purely a structural split. Test count
54 → 53 failures (1 pre-existing flaky test happened to pass), well
under baseline.

Closes #126.
