+++ src/cli.ts
@@ -8,14 +8,17 @@ import { generateBashCompletions, installBashCompletions } from "./completions";
 import type { RegistryConfigEntry } from "./config";
 import { DEFAULT_CONFIG, getConfigPath, loadConfig, loadUserConfig, saveConfig } from "./config";
 import { getConfigValue, listConfig, setConfigValue, unsetConfigValue } from "./config-cli";
+import { akmCurate } from "./curate";
 import { closeDatabase, openDatabase } from "./db";
 import { ConfigError, NotFoundError, UsageError } from "./errors";
-import { akmIndex, type IndexResponse } from "./indexer";
+import { akmIndex } from "./indexer";
 import { assembleInfo } from "./info";
 import { akmInit } from "./init";
-import { formatInstallAuditSummary } from "./install-audit";
 import { akmListSources, akmRemove, akmUpdate } from "./installed-stashes";
 import { renderMigrationHelp } from "./migration-help";
+import { getOutputMode, initOutputMode, type OutputMode, parseFlagValue } from "./output-context";
+import { shapeForCommand } from "./output-shapes";
+import { formatPlain, outputJsonl } from "./output-text";
 import { getCacheDir, getDbPath, getDefaultStashDir } from "./paths";
 import { buildRegistryIndex, writeRegistryIndex } from "./registry-build-index";
 import { searchRegistry } from "./registry-search";
@@ -27,15 +30,7 @@ import { parseAssetRef } from "./stash-ref";
 import { akmSearch, parseSearchSource } from "./stash-search";
 import { akmShowUnified } from "./stash-show";
 import { addStash } from "./stash-source-manage";
-import type {
-  KnowledgeView,
-  RegistrySearchResultHit,
-  SearchResponse,
-  ShowDetailLevel,
-  ShowResponse,
-  SourceKind,
-  StashSearchHit,
-} from "./stash-types";
+import type { KnowledgeView, ShowDetailLevel, SourceKind } from "./stash-types";
 import { insertUsageEvent } from "./usage-events";
 import { pkgVersion } from "./version";
 import { setQuiet, warn } from "./warn";
@@ -55,18 +50,6 @@ import {
   startWorkflowRun,
 } from "./workflow-runs";
 
-type OutputFormat = "json" | "yaml" | "text" | "jsonl";
-type DetailLevel = "brief" | "normal" | "full" | "summary";
-
-interface OutputMode {
-  format: OutputFormat;
-  detail: DetailLevel;
-  forAgent: boolean;
-}
-
-const OUTPUT_FORMATS: OutputFormat[] = ["json", "yaml", "text", "jsonl"];
-const DETAIL_LEVELS: DetailLevel[] = ["brief", "normal", "full", "summary"];
-const NORMAL_DESCRIPTION_LIMIT = 250;
 const MAX_CAPTURED_ASSET_SLUG_LENGTH = 64;
 const CONTEXT_HUB_ALIAS_REF = "context-hub";
 const CONTEXT_HUB_ALIAS_URL = "https://github.com/andrewyng/context-hub";
@@ -76,43 +59,8 @@ const SKILLS_SH_PROVIDER = "skills-sh";
 
 import { stringify as yamlStringify } from "yaml";
 
-function parseOutputFormat(value: string | undefined): OutputFormat | undefined {
-  if (!value) return undefined;
-  if ((OUTPUT_FORMATS as string[]).includes(value)) return value as OutputFormat;
-  throw new UsageError(`Invalid value for --format: ${value}. Expected one of: ${OUTPUT_FORMATS.join("|")}`);
-}
-
-function parseDetailLevel(value: string | undefined): DetailLevel | undefined {
-  if (!value) return undefined;
-  if ((DETAIL_LEVELS as string[]).includes(value)) return value as DetailLevel;
-  throw new UsageError(`Invalid value for --detail: ${value}. Expected one of: ${DETAIL_LEVELS.join("|")}`);
-}
-
-function parseFlagValue(flag: string): string | undefined {
-  for (let i = 0; i < process.argv.length; i++) {
-    const arg = process.argv[i];
-    if (arg === flag) return process.argv[i + 1];
-    if (arg.startsWith(`${flag}=`)) return arg.slice(flag.length + 1);
-  }
-  return undefined;
-}
-
-// Uses process.argv directly because the global output() function (called by all
-// commands) needs this flag but doesn't have access to citty's parsed args.
-function hasBooleanFlag(flag: string): boolean {
-  return process.argv.some((arg) => arg === flag || arg === `${flag}=true`);
-}
-
-function resolveOutputMode(): OutputMode {
-  const config = loadConfig();
-  const format = parseOutputFormat(parseFlagValue("--format")) ?? config.output?.format ?? "json";
-  const detail = parseDetailLevel(parseFlagValue("--detail")) ?? config.output?.detail ?? "brief";
-  const forAgent = hasBooleanFlag("--for-agent");
-  return { format, detail, forAgent };
-}
-
 function output(command: string, result: unknown): void {
-  const mode = resolveOutputMode();
+  const mode: OutputMode = getOutputMode();
   const shaped = shapeForCommand(command, result, mode.detail, mode.forAgent);
 
   if (mode.format === "jsonl") {
@@ -134,1010 +82,6 @@ function output(command: string, result: unknown): void {
     }
   }
 }
-
-function outputJsonl(command: string, shaped: unknown): void {
-  if (command === "search" || command === "registry-search") {
-    const r = shaped as Record<string, unknown>;
-    const hits = Array.isArray(r.hits) ? (r.hits as Record<string, unknown>[]) : [];
-    for (const hit of hits) {
-      console.log(JSON.stringify(hit));
-    }
-    const registryHits = Array.isArray(r.registryHits) ? (r.registryHits as Record<string, unknown>[]) : [];
-    for (const hit of registryHits) {
-      console.log(JSON.stringify(hit));
-    }
-    return;
-  }
-  // For non-search commands, output the whole object as a single JSONL line
-  console.log(JSON.stringify(shaped));
-}
-
-function shapeForCommand(command: string, result: unknown, detail: DetailLevel, forAgent = false): unknown {
-  switch (command) {
-    case "search":
-      return shapeSearchOutput(result as Record<string, unknown>, detail, forAgent);
-    case "registry-search":
-      return shapeRegistrySearchOutput(result as Record<string, unknown>, detail);
-    case "show":
-      return shapeShowOutput(result as Record<string, unknown>, detail, forAgent);
-    default:
-      return result;
-  }
-}
-
-function shapeSearchOutput(
-  result: Record<string, unknown>,
-  detail: DetailLevel,
-  forAgent = false,
-): Record<string, unknown> {
-  const hits = Array.isArray(result.hits) ? (result.hits as Record<string, unknown>[]) : [];
-  const registryHits = Array.isArray(result.registryHits) ? (result.registryHits as Record<string, unknown>[]) : [];
-  const shapedHits = forAgent
-    ? hits.map((hit) => shapeSearchHitForAgent(hit))
-    : hits.map((hit) => shapeSearchHit(hit, detail));
-  const shapedRegistryHits = forAgent
-    ? registryHits.map((hit) => shapeSearchHitForAgent(hit))
-    : registryHits.map((hit) => shapeSearchHit(hit, detail));
-
-  if (forAgent) {
-    return {
-      hits: shapedHits,
-      ...(shapedRegistryHits.length > 0 ? { registryHits: shapedRegistryHits } : {}),
-      ...(result.tip ? { tip: result.tip } : {}),
-    };
-  }
-
-  if (detail === "full") {
-    return {
-      schemaVersion: result.schemaVersion,
-      stashDir: result.stashDir,
-      source: result.source,
-      hits: shapedHits,
-      ...(shapedRegistryHits.length > 0 ? { registryHits: shapedRegistryHits } : {}),
-      ...(result.semanticSearch ? { semanticSearch: result.semanticSearch } : {}),
-      ...(result.tip ? { tip: result.tip } : {}),
-      ...(result.warnings ? { warnings: result.warnings } : {}),
-      ...(result.timing ? { timing: result.timing } : {}),
-    };
-  }
-
-  return {
-    hits: shapedHits,
-    ...(shapedRegistryHits.length > 0 ? { registryHits: shapedRegistryHits } : {}),
-    ...(Array.isArray(result.warnings) && result.warnings.length > 0 ? { warnings: result.warnings } : {}),
-    ...(result.tip ? { tip: result.tip } : {}),
-  };
-}
-
-function shapeRegistrySearchOutput(result: Record<string, unknown>, detail: DetailLevel): Record<string, unknown> {
-  const hits = Array.isArray(result.hits) ? (result.hits as Record<string, unknown>[]) : [];
-  const assetHits = Array.isArray(result.assetHits) ? (result.assetHits as Record<string, unknown>[]) : [];
-
-  // Shape stash hits as registry type
-  const shapedKitHits = hits.map((hit) => shapeSearchHit({ ...hit, type: "registry" }, detail));
-
-  // Shape asset hits by detail level
-  const shapedAssetHits = assetHits.map((hit) => shapeAssetHit(hit, detail));
-
-  const shaped: Record<string, unknown> = {
-    hits: shapedKitHits,
-    ...(shapedAssetHits.length > 0 ? { assetHits: shapedAssetHits } : {}),
-    ...(Array.isArray(result.warnings) && result.warnings.length > 0 ? { warnings: result.warnings } : {}),
-  };
-
-  if (detail === "full") {
-    shaped.query = result.query;
-  }
-
-  return shaped;
-}
-
-function shapeAssetHit(hit: Record<string, unknown>, detail: DetailLevel): Record<string, unknown> {
-  if (detail === "brief") return pickFields(hit, ["assetName", "assetType", "action", "estimatedTokens"]);
-  if (detail === "normal") {
-    return capDescription(
-      pickFields(hit, ["assetName", "assetType", "description", "stash", "action", "estimatedTokens"]),
-      NORMAL_DESCRIPTION_LIMIT,
-    );
-  }
-  return hit;
-}
-
-function shapeSearchHit(hit: Record<string, unknown>, detail: DetailLevel): Record<string, unknown> {
-  if (hit.type === "registry") {
-    if (detail === "brief") return pickFields(hit, ["name", "action"]);
-    if (detail === "normal") {
-      return capDescription(pickFields(hit, ["name", "description", "action", "curated"]), NORMAL_DESCRIPTION_LIMIT);
-    }
-    return hit;
-  }
-
-  // Stash hit (local or remote)
-  if (detail === "brief") return pickFields(hit, ["type", "name", "action", "estimatedTokens"]);
-  if (detail === "normal") {
-    return capDescription(
-      pickFields(hit, ["type", "name", "description", "action", "score", "estimatedTokens"]),
-      NORMAL_DESCRIPTION_LIMIT,
-    );
-  }
-  return hit;
-}
-
-/** Agent-optimized search hit: only fields an LLM agent needs to decide and act */
-function shapeSearchHitForAgent(hit: Record<string, unknown>): Record<string, unknown> {
-  const picked = pickFields(hit, ["name", "ref", "type", "description", "action", "score", "estimatedTokens"]);
-  return capDescription(picked, NORMAL_DESCRIPTION_LIMIT);
-}
-
-function capDescription(hit: Record<string, unknown>, limit: number): Record<string, unknown> {
-  if (typeof hit.description !== "string") return hit;
-  return { ...hit, description: truncateDescription(hit.description, limit) };
-}
-
-function truncateDescription(description: string, limit: number): string {
-  const normalized = description.replace(/\s+/g, " ").trim();
-  if (normalized.length <= limit) return normalized;
-
-  const truncated = normalized.slice(0, limit - 1);
-  const lastSpace = truncated.lastIndexOf(" ");
-  const safe = lastSpace >= Math.floor(limit * 0.6) ? truncated.slice(0, lastSpace) : truncated;
-  return `${safe.trimEnd()}...`;
-}
-
-function shapeShowOutput(
-  result: Record<string, unknown>,
-  detail: DetailLevel,
-  forAgent = false,
-): Record<string, unknown> {
-  if (forAgent) {
-    return pickFields(result, [
-      "type",
-      "name",
-      "description",
-      "action",
-      "content",
-      "template",
-      "prompt",
-      "run",
-      "setup",
-      "cwd",
-      "toolPolicy",
-      "modelHint",
-      "agent",
-      "parameters",
-      "workflowTitle",
-      "workflowParameters",
-      "steps",
-      "keys",
-      "comments",
-    ]);
-  }
-  if (detail === "summary") {
-    return pickFields(result, [
-      "type",
-      "name",
-      "description",
-      "tags",
-      "parameters",
-      "workflowTitle",
-      "action",
-      "run",
-      "origin",
-      "keys",
-      "comments",
-    ]);
-  }
-
-  const base = pickFields(result, [
-    "type",
-    "name",
-    "origin",
-    "action",
-    "description",
-    "tags",
-    "content",
-    "template",
-    "prompt",
-    "toolPolicy",
-    "modelHint",
-    "agent",
-    "parameters",
-    "workflowTitle",
-    "workflowParameters",
-    "steps",
-    "run",
-    "setup",
-    "cwd",
-    "keys",
-    "comments",
-  ]);
-
-  if (detail !== "full") {
-    return base;
-  }
-
-  return {
-    schemaVersion: 1,
-    ...base,
-    ...pickFields(result, ["path", "editable", "editHint"]),
-  };
-}
-
-function pickFields(source: Record<string, unknown>, fields: string[]): Record<string, unknown> {
-  const result: Record<string, unknown> = {};
-  for (const field of fields) {
-    if (source[field] !== undefined) {
-      result[field] = source[field];
-    }
-  }
-  return result;
-}
-
-/**
- * Return a plain-text string for commands that are better as short messages,
- * or null to fall through to YAML output.
- */
-function formatPlain(command: string, result: unknown, detail: DetailLevel): string | null {
-  const r = result as Record<string, unknown>;
-
-  switch (command) {
-    case "init": {
-      let out = `Stash initialized at ${r.stashDir ?? "unknown"}`;
-      if (r.configPath) out += `\nConfig saved to ${r.configPath}`;
-      return out;
-    }
-    case "index": {
-      const indexResult = result as Partial<IndexResponse>;
-      let out = `Indexed ${indexResult.totalEntries ?? 0} entries from ${indexResult.directoriesScanned ?? 0} directories (mode: ${indexResult.mode ?? "unknown"})`;
-      const warnings = indexResult.warnings;
-      if (Array.isArray(warnings) && warnings.length > 0) {
-        out += `\nWarnings (${warnings.length}):`;
-        for (const message of warnings) out += `\n  - ${String(message)}`;
-      }
-      const verification = indexResult.verification;
-      if (verification?.ok === false && verification.message) {
-        out += `\nVerification: ${String(verification.message)}`;
-      }
-      return out;
-    }
-    case "show": {
-      const lines: string[] = [];
-      if (r.type || r.name) {
-        lines.push(`# ${String(r.type ?? "asset")}: ${String(r.name ?? "unknown")}`);
-      }
-      if (r.origin !== undefined) lines.push(`# origin: ${String(r.origin)}`);
-      if (r.action) lines.push(`# ${String(r.action)}`);
-      if (r.description) lines.push(`description: ${String(r.description)}`);
-      if (r.workflowTitle) lines.push(`workflowTitle: ${String(r.workflowTitle)}`);
-      if (r.agent) lines.push(`agent: ${String(r.agent)}`);
-      if (Array.isArray(r.parameters) && r.parameters.length > 0) lines.push(`parameters: ${r.parameters.join(", ")}`);
-      if (Array.isArray(r.workflowParameters) && r.workflowParameters.length > 0) {
-        lines.push("workflowParameters:");
-        for (const parameter of r.workflowParameters as Array<Record<string, unknown>>) {
-          const name = typeof parameter.name === "string" ? parameter.name : "unknown";
-          const description =
-            typeof parameter.description === "string" && parameter.description.trim()
-              ? `: ${parameter.description}`
-              : "";
-          lines.push(`  - ${name}${description}`);
-        }
-      }
-      if (r.modelHint != null) lines.push(`modelHint: ${String(r.modelHint)}`);
-      if (r.toolPolicy != null) lines.push(`toolPolicy: ${JSON.stringify(r.toolPolicy)}`);
-      if (r.run) lines.push(`run: ${String(r.run)}`);
-      if (r.setup) lines.push(`setup: ${String(r.setup)}`);
-      if (r.cwd) lines.push(`cwd: ${String(r.cwd)}`);
-      if (detail === "full") {
-        if (r.path) lines.push(`path: ${String(r.path)}`);
-        if (r.editable !== undefined) lines.push(`editable: ${String(r.editable)}`);
-        if (r.editHint) lines.push(`editHint: ${String(r.editHint)}`);
-        if (r.schemaVersion !== undefined) lines.push(`schemaVersion: ${String(r.schemaVersion)}`);
-      }
-      const payloads = [r.content, r.template, r.prompt].filter((value) => value != null).map(String);
-      if (Array.isArray(r.steps) && r.steps.length > 0) {
-        if (lines.length > 0) lines.push("");
-        lines.push("steps:");
-        for (const [index, step] of (r.steps as Array<Record<string, unknown>>).entries()) {
-          const title = typeof step.title === "string" ? step.title : "Untitled step";
-          const id = typeof step.id === "string" ? step.id : "unknown";
-          lines.push(`  ${index + 1}. ${title} [${id}]`);
-          if (typeof step.instructions === "string" && step.instructions.trim()) {
-            lines.push(`     instructions: ${step.instructions.replace(/\n+/g, " ").trim()}`);
-          }
-          if (Array.isArray(step.completionCriteria) && step.completionCriteria.length > 0) {
-            lines.push("     completion:");
-            for (const criterion of step.completionCriteria) {
-              lines.push(`       - ${String(criterion)}`);
-            }
-          }
-        }
-      }
-      if (payloads.length > 0) {
-        if (lines.length > 0) lines.push("");
-        lines.push(...payloads);
-      }
-      return lines.length > 0 ? lines.join("\n") : null;
-    }
-    case "search": {
-      return formatSearchPlain(r, detail);
-    }
-    case "curate": {
-      return formatCuratePlain(r, detail);
-    }
-    case "wiki-list": {
-      return formatWikiListPlain(r);
-    }
-    case "wiki-show": {
-      return formatWikiShowPlain(r);
-    }
-    case "wiki-create": {
-      return formatWikiCreatePlain(r);
-    }
-    case "wiki-remove": {
-      return formatWikiRemovePlain(r);
-    }
-    case "wiki-pages": {
-      return formatWikiPagesPlain(r);
-    }
-    case "wiki-stash": {
-      return formatWikiStashPlain(r);
-    }
-    case "wiki-lint": {
-      return formatWikiLintPlain(r);
-    }
-    case "wiki-ingest": {
-      return formatWikiIngestPlain(r);
-    }
-    case "workflow-start":
-    case "workflow-status":
-    case "workflow-complete": {
-      return formatWorkflowStatusPlain(r);
-    }
-    case "workflow-next": {
-      return formatWorkflowNextPlain(r);
-    }
-    case "workflow-list": {
-      return formatWorkflowListPlain(r);
-    }
-    case "workflow-create": {
-      if (r.ref && r.path) {
-        return `Created ${String(r.ref)} at ${String(r.path)}`;
-      }
-      return null;
-    }
-    case "list": {
-      const sources = Array.isArray(r.sources) ? (r.sources as Record<string, unknown>[]) : [];
-      if (sources.length === 0) return "No sources configured. Use `akm add` to add a source.";
-      const lines: string[] = [];
-      for (const src of sources) {
-        const kind = typeof src.kind === "string" ? src.kind : "unknown";
-        const name = typeof src.name === "string" ? src.name : "unnamed";
-        const ver = typeof src.version === "string" ? ` v${src.version}` : "";
-        const prov = typeof src.provider === "string" ? ` (${src.provider})` : "";
-        const flags: string[] = [];
-        if (typeof src.wiki === "string") flags.push(`wiki:${src.wiki}`);
-        if (src.updatable === true) flags.push("updatable");
-        if (src.writable === true) flags.push("writable");
-        const flagText = flags.length > 0 ? ` [${flags.join(", ")}]` : "";
-        lines.push(`[${kind}] ${name}${ver}${prov}${flagText}`);
-      }
-      return lines.join("\n");
-    }
-    case "add": {
-      const index = r.index as Record<string, unknown> | undefined;
-      const scanned = index?.directoriesScanned ?? 0;
-      const total = index?.totalEntries ?? 0;
-      const lines = [`Installed ${r.ref} (${scanned} directories scanned, ${total} total assets indexed)`];
-      const warnings = index?.warnings;
-      if (Array.isArray(warnings) && warnings.length > 0) {
-        lines.push(`Warnings (${warnings.length}):`);
-        for (const message of warnings) lines.push(`  - ${String(message)}`);
-      }
-      const installed = r.installed as Record<string, unknown> | undefined;
-      const audit = installed?.audit;
-      if (audit && typeof audit === "object") {
-        lines.push(formatInstallAuditSummary(audit as Parameters<typeof formatInstallAuditSummary>[0]));
-      }
-      return lines.join("\n");
-    }
-    case "remove": {
-      const target = r.target ?? r.ref ?? "";
-      const ok = r.ok !== false ? "OK" : "FAILED";
-      return `remove: ${target} ${ok}`;
-    }
-    case "update": {
-      const processed = r.processed as Array<Record<string, unknown>> | undefined;
-      if (!processed?.length) return `update: nothing to update`;
-      const lines = processed.map((item) => {
-        const changed = item.changed as Record<string, unknown> | undefined;
-        const installed = item.installed as Record<string, unknown> | undefined;
-        const previous = item.previous as Record<string, unknown> | undefined;
-        if (changed?.any) {
-          const prev = previous?.resolvedVersion ?? "unknown";
-          const next = installed?.resolvedVersion ?? "unknown";
-          return `update: ${item.id} v${prev} → v${next}`;
-        }
-        return `update: ${item.id} (unchanged)`;
-      });
-      return lines.join("\n");
-    }
-    case "upgrade": {
-      if (r.upgraded === true) {
-        return `akm upgraded: v${r.currentVersion} → v${r.newVersion}`;
-      }
-      if (r.updateAvailable === true) {
-        return `akm v${r.currentVersion} → v${r.latestVersion} available (run 'akm upgrade' to install)`;
-      }
-      if (r.updateAvailable === false && r.latestVersion) {
-        return `akm v${r.currentVersion} is already the latest version`;
-      }
-      if (r.message) return String(r.message);
-      return null;
-    }
-    case "clone": {
-      const dst = (r.destination as Record<string, unknown>)?.path ?? "unknown";
-      const remote = r.remoteFetched ? " (fetched from remote)" : "";
-      const over = r.overwritten ? " (overwritten)" : "";
-      return `Cloned${remote} → ${dst}${over}`;
-    }
-    default:
-      return null; // fall through to YAML
-  }
-}
-
-function formatWorkflowListPlain(result: Record<string, unknown>): string {
-  const runs = Array.isArray(result.runs) ? (result.runs as Array<Record<string, unknown>>) : [];
-  if (runs.length === 0) return "No workflow runs found.";
-
-  return runs
-    .map((run) => {
-      const id = typeof run.id === "string" ? run.id : "unknown";
-      const ref = typeof run.workflowRef === "string" ? run.workflowRef : "workflow:unknown";
-      const status = typeof run.status === "string" ? run.status : "unknown";
-      const currentStep = typeof run.currentStepId === "string" ? ` (current: ${run.currentStepId})` : "";
-      return `${id} ${ref} [${status}]${currentStep}`;
-    })
-    .join("\n");
-}
-
-function formatWorkflowStatusPlain(result: Record<string, unknown>): string | null {
-  const run =
-    typeof result.run === "object" && result.run !== null ? (result.run as Record<string, unknown>) : undefined;
-  const workflow =
-    typeof result.workflow === "object" && result.workflow !== null
-      ? (result.workflow as Record<string, unknown>)
-      : undefined;
-  if (!run || !workflow) return null;
-
-  const lines = [
-    `workflow: ${String(workflow.ref ?? "workflow:unknown")}`,
-    `run: ${String(run.id ?? "unknown")}`,
-    `title: ${String(run.workflowTitle ?? workflow.title ?? "Workflow")}`,
-    `status: ${String(run.status ?? "unknown")}`,
-  ];
-  if (run.currentStepId) lines.push(`currentStep: ${String(run.currentStepId)}`);
-
-  const steps = Array.isArray(workflow.steps) ? (workflow.steps as Array<Record<string, unknown>>) : [];
-  if (steps.length > 0) {
-    lines.push("steps:");
-    for (const step of steps) {
-      const title = typeof step.title === "string" ? step.title : "Untitled step";
-      const id = typeof step.id === "string" ? step.id : "unknown";
-      const status = typeof step.status === "string" ? step.status : "unknown";
-      lines.push(`  - ${title} [${id}] (${status})`);
-      if (typeof step.notes === "string" && step.notes.trim()) {
-        lines.push(`    notes: ${step.notes}`);
-      }
-    }
-  }
-  return lines.join("\n");
-}
-
-function formatWorkflowNextPlain(result: Record<string, unknown>): string | null {
-  const base = formatWorkflowStatusPlain(result);
-  const step =
-    typeof result.step === "object" && result.step !== null ? (result.step as Record<string, unknown>) : undefined;
-  if (!step) return base;
-
-  const lines = base ? [base, "", "next:"] : ["next:"];
-  lines.push(`  ${String(step.title ?? "Untitled step")} [${String(step.id ?? "unknown")}]`);
-  if (typeof step.instructions === "string" && step.instructions.trim()) {
-    lines.push(`  instructions: ${step.instructions.replace(/\n+/g, " ").trim()}`);
-  }
-  const completion = Array.isArray(step.completionCriteria) ? step.completionCriteria : [];
-  if (completion.length > 0) {
-    lines.push("  completion:");
-    for (const criterion of completion) {
-      lines.push(`    - ${String(criterion)}`);
-    }
-  }
-  return lines.join("\n");
-}
-
-function formatSearchPlain(r: Record<string, unknown>, detail: DetailLevel): string {
-  const hits = (r.hits as Record<string, unknown>[]) ?? [];
-  const registryHits = (r.registryHits as Record<string, unknown>[]) ?? [];
-  const allHits = [...hits, ...registryHits];
-
-  if (allHits.length === 0) {
-    return r.tip ? String(r.tip) : "No results found.";
-  }
-
-  const lines: string[] = [];
-
-  for (const hit of allHits) {
-    const type = hit.type ?? "unknown";
-    const name = hit.name ?? "unnamed";
-    const score = hit.score != null ? ` (score: ${hit.score})` : "";
-    const desc = hit.description ? `  ${hit.description}` : "";
-
-    lines.push(`${type}: ${name}${score}`);
-    if (desc) lines.push(desc);
-
-    if (hit.id) lines.push(`  id: ${String(hit.id)}`);
-    if (hit.ref) lines.push(`  ref: ${String(hit.ref)}`);
-    if (hit.origin !== undefined) lines.push(`  origin: ${String(hit.origin)}`);
-    if (hit.size) lines.push(`  size: ${String(hit.size)}`);
-    if (hit.action) lines.push(`  action: ${String(hit.action)}`);
-    if (hit.run) lines.push(`  run: ${String(hit.run)}`);
-    if (Array.isArray(hit.tags) && hit.tags.length > 0) lines.push(`  tags: ${hit.tags.join(", ")}`);
-    if (hit.curated !== undefined) lines.push(`  curated: ${String(hit.curated)}`);
-
-    if (detail === "full") {
-      if (hit.path) lines.push(`  path: ${String(hit.path)}`);
-      if (hit.editable != null) lines.push(`  editable: ${String(hit.editable)}`);
-      if (hit.editHint) lines.push(`  editHint: ${String(hit.editHint)}`);
-      const whyMatched = hit.whyMatched as string[] | undefined;
-      if (whyMatched && whyMatched.length > 0) {
-        lines.push(`  whyMatched: ${whyMatched.join(", ")}`);
-      }
-    }
-
-    lines.push(""); // blank line between hits
-  }
-
-  if (detail === "full" && r.timing) {
-    const timing = r.timing as Record<string, unknown>;
-    const parts: string[] = [];
-    if (timing.totalMs != null) parts.push(`total: ${timing.totalMs}ms`);
-    if (timing.rankMs != null) parts.push(`rank: ${timing.rankMs}ms`);
-    if (timing.embedMs != null) parts.push(`embed: ${timing.embedMs}ms`);
-    if (parts.length > 0) lines.push(`timing: ${parts.join(", ")}`);
-  }
-
-  return lines.join("\n").trimEnd();
-}
-
-function formatWikiListPlain(r: Record<string, unknown>): string {
-  const wikis = Array.isArray(r.wikis) ? (r.wikis as Array<Record<string, unknown>>) : [];
-  if (wikis.length === 0)
-    return "No wikis. Create one with `akm wiki create <name>` or register one with `akm wiki register <name> <path-or-repo>`.";
-  const lines = ["NAME\tPAGES\tRAWS\tLAST-MODIFIED"];
-  for (const w of wikis) {
-    const name = typeof w.name === "string" ? w.name : "?";
-    const pages = typeof w.pages === "number" ? w.pages : 0;
-    const raws = typeof w.raws === "number" ? w.raws : 0;
-    const modified = typeof w.lastModified === "string" ? w.lastModified : "-";
-    lines.push(`${name}\t${pages}\t${raws}\t${modified}`);
-  }
-  return lines.join("\n");
-}
-
-function formatWikiShowPlain(r: Record<string, unknown>): string {
-  const lines: string[] = [];
-  if (r.name) lines.push(`# wiki: ${String(r.name)}`);
-  if (r.path) lines.push(`path: ${String(r.path)}`);
-  if (r.description) lines.push(`description: ${String(r.description)}`);
-  if (typeof r.pages === "number") lines.push(`pages: ${r.pages}`);
-  if (typeof r.raws === "number") lines.push(`raws: ${r.raws}`);
-  if (r.lastModified) lines.push(`lastModified: ${String(r.lastModified)}`);
-  const recentLog = Array.isArray(r.recentLog) ? (r.recentLog as string[]) : [];
-  if (recentLog.length > 0) {
-    lines.push("", "recent log:");
-    for (const entry of recentLog) {
-      lines.push(entry);
-      lines.push("");
-    }
-  }
-  return lines.join("\n").trimEnd();
-}
-
-function formatWikiCreatePlain(r: Record<string, unknown>): string {
-  const created = Array.isArray(r.created) ? (r.created as string[]) : [];
-  const skipped = Array.isArray(r.skipped) ? (r.skipped as string[]) : [];
-  const lines = [`Created wiki ${String(r.ref ?? r.name)} at ${String(r.path ?? "?")}`];
-  if (created.length > 0) lines.push(`  created: ${created.length} file(s)`);
-  if (skipped.length > 0) lines.push(`  skipped: ${skipped.length} existing file(s)`);
-  return lines.join("\n");
-}
-
-function formatWikiRemovePlain(r: Record<string, unknown>): string {
-  const preserved = r.preservedRaw === true;
-  const removed = Array.isArray(r.removed) ? (r.removed as string[]).length : 0;
-  const base = `Removed wiki ${String(r.name ?? "?")} (${removed} path(s))`;
-  return preserved ? `${base}; preserved ${String(r.rawPath ?? "raw/")}` : base;
-}
-
-function formatWikiPagesPlain(r: Record<string, unknown>): string {
-  const pages = Array.isArray(r.pages) ? (r.pages as Array<Record<string, unknown>>) : [];
-  if (pages.length === 0) return `No pages in wiki:${String(r.wiki ?? "?")}.`;
-  const lines: string[] = [];
-  for (const p of pages) {
-    const ref = String(p.ref ?? "?");
-    const kind = typeof p.pageKind === "string" ? ` [${p.pageKind}]` : "";
-    const desc = typeof p.description === "string" && p.description ? ` — ${p.description}` : "";
-    lines.push(`${ref}${kind}${desc}`);
-  }
-  return lines.join("\n");
-}
-
-function formatWikiStashPlain(r: Record<string, unknown>): string {
-  const slug = String(r.slug ?? "?");
-  const pathValue = String(r.path ?? "?");
-  return `Stashed ${slug} → ${pathValue}`;
-}
-
-function formatWikiLintPlain(r: Record<string, unknown>): string {
-  const findings = Array.isArray(r.findings) ? (r.findings as Array<Record<string, unknown>>) : [];
-  const pagesScanned = typeof r.pagesScanned === "number" ? r.pagesScanned : 0;
-  const rawsScanned = typeof r.rawsScanned === "number" ? r.rawsScanned : 0;
-  const header = `${findings.length} finding(s) in wiki:${String(r.wiki ?? "?")} (${pagesScanned} page(s), ${rawsScanned} raw(s))`;
-  if (findings.length === 0) return `${header} — clean.`;
-  const lines = [header];
-  for (const f of findings) {
-    const kind = String(f.kind ?? "?");
-    const message = String(f.message ?? "");
-    lines.push(`- [${kind}] ${message}`);
-  }
-  return lines.join("\n");
-}
-
-function formatWikiIngestPlain(r: Record<string, unknown>): string {
-  if (typeof r.workflow === "string") return r.workflow;
-  return JSON.stringify(r, null, 2);
-}
-
-function formatCuratePlain(r: Record<string, unknown>, detail: DetailLevel): string {
-  const query = typeof r.query === "string" ? r.query : "";
-  const summary = typeof r.summary === "string" ? r.summary : "";
-  const items = Array.isArray(r.items) ? (r.items as Record<string, unknown>[]) : [];
-
-  const lines: string[] = [`Curated results for "${query}"`];
-  if (summary) lines.push(summary);
-  if (items.length === 0) {
-    if (r.tip) lines.push(String(r.tip));
-    return lines.join("\n");
-  }
-
-  for (const item of items) {
-    const type = typeof item.type === "string" ? item.type : "unknown";
-    const name = typeof item.name === "string" ? item.name : "unnamed";
-    lines.push("");
-    lines.push(`[${type}] ${name}`);
-    if (item.description) lines.push(`  ${String(item.description)}`);
-    if (item.preview) lines.push(`  preview: ${String(item.preview)}`);
-    if (item.ref) lines.push(`  ref: ${String(item.ref)}`);
-    if (item.id) lines.push(`  id: ${String(item.id)}`);
-    if (Array.isArray(item.parameters) && item.parameters.length > 0) {
-      lines.push(`  parameters: ${item.parameters.join(", ")}`);
-    }
-    if (item.run) lines.push(`  run: ${String(item.run)}`);
-    if (item.followUp) lines.push(`  show: ${String(item.followUp)}`);
-    if (detail !== "brief" && item.reason) lines.push(`  why: ${String(item.reason)}`);
-  }
-
-  const warnings = Array.isArray(r.warnings) ? r.warnings : [];
-  if (warnings.length > 0) {
-    lines.push("");
-    lines.push("Warnings:");
-    for (const warning of warnings) {
-      lines.push(`- ${String(warning)}`);
-    }
-  }
-
-  return lines.join("\n");
-}
-
-type CuratedStashItem = {
-  source: "stash";
-  type: string;
-  name: string;
-  ref: string;
-  description?: string;
-  preview?: string;
-  parameters?: string[];
-  run?: string;
-  followUp: string;
-  reason: string;
-  score?: number;
-};
-
-type CuratedRegistryItem = {
-  source: "registry";
-  type: "registry";
-  name: string;
-  id: string;
-  description?: string;
-  followUp: string;
-  reason: string;
-  score?: number;
-};
-
-type CuratedItem = CuratedStashItem | CuratedRegistryItem;
-
-const CURATE_FALLBACK_FILTER_WORDS = new Set([
-  "a",
-  "an",
-  "and",
-  "for",
-  "how",
-  "i",
-  "in",
-  "of",
-  "or",
-  "the",
-  "to",
-  "with",
-]);
-const CURATED_TYPE_FALLBACK_ORDER = ["skill", "command", "script", "knowledge", "agent", "memory"];
-const CURATED_TYPE_FALLBACK_INDEX = new Map(CURATED_TYPE_FALLBACK_ORDER.map((type, index) => [type, index]));
-const MIN_CURATE_FALLBACK_TOKEN_LENGTH = 3;
-const MAX_CURATE_FALLBACK_KEYWORDS = 6;
-const CURATE_SEARCH_LIMIT_MULTIPLIER = 4;
-const MIN_CURATE_SEARCH_LIMIT = 12;
-
-async function curateSearchResults(
-  query: string,
-  result: SearchResponse,
-  limit: number,
-  selectedType?: string,
-): Promise<{
-  query: string;
-  summary: string;
-  items: CuratedItem[];
-  warnings?: string[];
-  tip?: string;
-}> {
-  const stashHits = result.hits.filter((hit): hit is StashSearchHit => hit.type !== "registry");
-  const registryHits = result.registryHits ?? [];
-
-  let selectedStashHits: StashSearchHit[];
-  if (selectedType && selectedType !== "any") {
-    selectedStashHits = stashHits.slice(0, limit);
-  } else {
-    const bestByType = new Map<string, StashSearchHit>();
-    for (const hit of stashHits) {
-      if (!bestByType.has(hit.type)) bestByType.set(hit.type, hit);
-    }
-    const orderedTypes = orderCuratedTypes(query, Array.from(bestByType.keys()));
-    selectedStashHits = orderedTypes
-      .map((type) => bestByType.get(type))
-      .filter((hit): hit is StashSearchHit => Boolean(hit));
-  }
-
-  const selectedRegistryHits =
-    selectedStashHits.length >= limit ? [] : registryHits.slice(0, Math.min(2, limit - selectedStashHits.length));
-
-  const items = [
-    ...(await Promise.all(selectedStashHits.slice(0, limit).map((hit) => enrichCuratedStashHit(query, hit)))),
-    ...selectedRegistryHits.map((hit) => buildCuratedRegistryItem(query, hit)),
-  ].slice(0, limit);
-
-  return {
-    query,
-    summary: buildCurateSummary(query, items),
-    items,
-    ...(result.warnings?.length ? { warnings: result.warnings } : {}),
-    ...(result.tip ? { tip: result.tip } : {}),
-  };
-}
-
-function orderCuratedTypes(query: string, types: string[]): string[] {
-  const lower = query.toLowerCase();
-  const boosts = new Map<string, number>();
-  const addBoost = (type: string, amount: number) => boosts.set(type, (boosts.get(type) ?? 0) + amount);
-
-  if (/(run|script|bash|shell|cli|execute|automation|deploy|build|test|lint)/.test(lower)) {
-    addBoost("script", 6);
-    addBoost("command", 4);
-  }
-  if (/(guide|docs?|readme|reference|how|explain|learn|why)/.test(lower)) {
-    addBoost("knowledge", 6);
-    addBoost("skill", 4);
-  }
-  if (/(agent|assistant|planner|review|analy[sz]e|architect|prompt)/.test(lower)) {
-    addBoost("agent", 6);
-    addBoost("skill", 3);
-  }
-  if (/(config|template|release|generate|command)/.test(lower)) {
-    addBoost("command", 5);
-  }
-  if (/(memory|context|recall|remember)/.test(lower)) {
-    addBoost("memory", 6);
-  }
-
-  return [...types].sort((a, b) => {
-    const boostDiff = (boosts.get(b) ?? 0) - (boosts.get(a) ?? 0);
-    if (boostDiff !== 0) return boostDiff;
-    return (
-      (CURATED_TYPE_FALLBACK_INDEX.get(a) ?? Number.MAX_SAFE_INTEGER) -
-      (CURATED_TYPE_FALLBACK_INDEX.get(b) ?? Number.MAX_SAFE_INTEGER)
-    );
-  });
-}
-
-async function enrichCuratedStashHit(query: string, hit: StashSearchHit): Promise<CuratedStashItem> {
-  let shown: ShowResponse | undefined;
-  try {
-    shown = await akmShowUnified({ ref: hit.ref });
-  } catch {
-    shown = undefined;
-  }
-
-  const description = shown?.description ?? hit.description;
-  const preview = buildCuratedPreview(shown, hit);
-  return {
-    source: "stash",
-    type: shown?.type ?? hit.type,
-    name: shown?.name ?? hit.name,
-    ref: hit.ref,
-    ...(description ? { description } : {}),
-    ...(preview ? { preview } : {}),
-    ...(shown?.parameters?.length ? { parameters: shown.parameters } : {}),
-    ...(shown?.run ? { run: shown.run } : {}),
-    followUp: `akm show ${hit.ref}`,
-    reason: buildCuratedReason(query, shown?.type ?? hit.type),
-    ...(hit.score !== undefined ? { score: hit.score } : {}),
-  };
-}
-
-function buildCuratedRegistryItem(query: string, hit: RegistrySearchResultHit): CuratedRegistryItem {
-  return {
-    source: "registry",
-    type: "registry",
-    name: hit.name,
-    id: hit.id,
-    ...(hit.description ? { description: hit.description } : {}),
-    followUp: hit.action ?? `akm add ${hit.id}`,
-    reason: `Useful external source to explore for ${query}.`,
-    ...(hit.score !== undefined ? { score: hit.score } : {}),
-  };
-}
-
-function firstNonEmpty(values: Array<string | undefined>): string | undefined {
-  return values.find((value) => typeof value === "string" && value.trim().length > 0);
-}
-
-function buildCuratedPreview(shown: ShowResponse | undefined, hit: StashSearchHit): string | undefined {
-  if (shown?.run) return truncateDescription(`run ${shown.run}`, 160);
-  const payload = firstNonEmpty([shown?.template, shown?.prompt, shown?.content, hit.description])
-    ?.replace(/\s+/g, " ")
-    .trim();
-  return payload ? truncateDescription(payload, 160) : undefined;
-}
-
-function buildCuratedReason(query: string, type: string): string {
-  switch (type) {
-    case "script":
-      return `Best runnable script match for "${query}".`;
-    case "command":
-      return `Best reusable command/template match for "${query}".`;
-    case "knowledge":
-      return `Best reference document match for "${query}".`;
-    case "skill":
-      return `Best instructions/workflow match for "${query}".`;
-    case "agent":
-      return `Best specialized agent prompt match for "${query}".`;
-    case "memory":
-      return `Best saved context match for "${query}".`;
-    default:
-      return `Best ${type} match for "${query}".`;
-  }
-}
-
-function buildCurateSummary(query: string, items: CuratedItem[]): string {
-  if (items.length === 0) {
-    return `No curated assets were selected for "${query}".`;
-  }
-  const labels = items.map((item) => `${item.type}:${item.name}`);
-  return `Selected ${items.length} high-signal result${items.length === 1 ? "" : "s"}: ${labels.join(", ")}.`;
-}
-
-function hasSearchResults(result: SearchResponse): boolean {
-  return result.hits.length > 0 || (result.registryHits?.length ?? 0) > 0;
-}
-
-/**
- * Extract a small set of fallback keywords when a prompt-style curate query
- * returns no hits as a whole phrase.
- *
- * We keep up to MAX_CURATE_FALLBACK_KEYWORDS distinct keywords and drop short
- * or common filler words so follow-up searches stay inexpensive while focusing
- * on higher-signal terms.
- */
-function deriveCurateFallbackQueries(query: string): string[] {
-  return Array.from(
-    new Set(
-      query
-        .toLowerCase()
-        .split(/[^a-z0-9]+/)
-        .map((token) => token.trim())
-        // Keep longer tokens so fallback stays focused on higher-signal terms
-        // and avoids broad one- and two-letter matches that overwhelm curation.
-        .filter(
-          (token) => token.length >= MIN_CURATE_FALLBACK_TOKEN_LENGTH && !CURATE_FALLBACK_FILTER_WORDS.has(token),
-        ),
-    ),
-  ).slice(0, MAX_CURATE_FALLBACK_KEYWORDS);
-}
-
-function mergeCurateSearchResponses(base: SearchResponse, extras: SearchResponse[]): SearchResponse {
-  const hitsByRef = new Map<string, StashSearchHit>();
-  for (const hit of base.hits.filter((entry): entry is StashSearchHit => entry.type !== "registry")) {
-    hitsByRef.set(hit.ref, hit);
-  }
-  for (const result of extras) {
-    for (const hit of result.hits.filter((entry): entry is StashSearchHit => entry.type !== "registry")) {
-      const existing = hitsByRef.get(hit.ref);
-      if (!existing || (hit.score ?? 0) > (existing.score ?? 0)) {
-        hitsByRef.set(hit.ref, hit);
-      }
-    }
-  }
-
-  const registryById = new Map<string, RegistrySearchResultHit>();
-  for (const hit of base.registryHits ?? []) {
-    registryById.set(hit.id, hit);
-  }
-  for (const result of extras) {
-    for (const hit of result.registryHits ?? []) {
-      const existing = registryById.get(hit.id);
-      if (!existing || (hit.score ?? 0) > (existing.score ?? 0)) {
-        registryById.set(hit.id, hit);
-      }
-    }
-  }
-
-  const warnings = Array.from(
-    new Set([...(base.warnings ?? []), ...extras.flatMap((result) => result.warnings ?? [])]),
-  );
-  const mergedHits = [...hitsByRef.values()].sort((a, b) => (b.score ?? 0) - (a.score ?? 0));
-  const mergedRegistryHits = [...registryById.values()].sort((a, b) => (b.score ?? 0) - (a.score ?? 0));
-
-  return {
-    ...base,
-    hits: mergedHits,
-    ...(mergedRegistryHits.length > 0 ? { registryHits: mergedRegistryHits } : {}),
-    ...(warnings.length > 0 ? { warnings } : {}),
-    ...(mergedHits.length > 0 || mergedRegistryHits.length > 0 ? { tip: undefined } : {}),
-  };
-}
-
-async function searchForCuration(input: {
-  query: string;
-  type?: string;
-  limit: number;
-  source: ReturnType<typeof parseSearchSource>;
-}): Promise<SearchResponse> {
-  const initial = await akmSearch(input);
-  if (hasSearchResults(initial)) return initial;
-
-  const fallbackQueries = deriveCurateFallbackQueries(input.query);
-  if (fallbackQueries.length <= 1) return initial;
-
-  const fallbackResults = await Promise.all(
-    fallbackQueries.map((token) =>
-      akmSearch({
-        query: token,
-        type: input.type,
-        limit: input.limit,
-        source: input.source,
-      }),
-    ),
-  );
-  return mergeCurateSearchResponses(initial, fallbackResults);
-}
-
 /**
  * Module Naming:
  * - stash-*          : Asset operations (search, show, add, clone)
@@ -1252,15 +196,7 @@ const curateCommand = defineCommand({
       }
       const limit = limitRaw && limitRaw > 0 ? limitRaw : 4;
       const source = parseSearchSource(args.source ?? "stash");
-      const searchResult = await searchForCuration({
-        query: args.query,
-        type,
-        // Search deeper than the final curated count so we can pick one strong
-        // match per type and still have room for fallback retries.
-        limit: Math.max(limit * CURATE_SEARCH_LIMIT_MULTIPLIER, MIN_CURATE_SEARCH_LIMIT),
-        source,
-      });
-      const curated = await curateSearchResults(args.query, searchResult, limit, type);
+      const curated = await akmCurate({ query: args.query, type, limit, source });
       output("curate", curated);
     });
   },
@@ -1523,7 +459,7 @@ const showCommand = defineCommand({
         }
       }
       // Map CLI detail level to ShowDetailLevel for the show function
-      const cliDetail = resolveOutputMode().detail;
+      const cliDetail = getOutputMode().detail;
       const showDetail: ShowDetailLevel | undefined = cliDetail === "summary" ? "summary" : undefined;
       const result = await akmShowUnified({ ref: args.ref, view, detail: showDetail });
       output("show", result);
@@ -1652,7 +588,7 @@ const saveCommand = defineCommand({
       // before any standalone occurrence of the same value in the save
       // subcommand's argv slice. This preserves legitimate invocations
       // like `akm save json --format json`.
-      const parsedFormat = parseFlagValue("--format");
+      const parsedFormat = parseFlagValue(process.argv, "--format");
       const effectiveName =
         args.name !== undefined &&
         parsedFormat !== undefined &&
@@ -3009,6 +1945,9 @@ const SHOW_VIEW_MODES = new Set(["toc", "frontmatter", "full", "section", "lines
 // citty reads process.argv directly and does not accept a custom argv array,
 // so we must replace process.argv with the normalized version before runMain.
 process.argv = normalizeShowArgv(process.argv);
+// Resolve output mode once at startup from the (normalized) argv and persisted
+// config. All subsequent output() calls read from this in-memory singleton.
+initOutputMode(process.argv, loadConfig().output ?? {});
 runMain(main);
 
 // ── Exit codes ──────────────────────────────────────────────────────────────
